+++ config/events-web.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 use Yiisoft\View\Event\WebView\AfterRender;
-use Yiisoft\Yii\View\Debug\WebViewCollector;
+use Yiisoft\Yii\View\Renderer\Debug\WebViewCollector;
 
 if (!($params['yiisoft/yii-debug']['enabled'] ?? false)) {
     return [];