+++ tests/Support/Controller/SubNamespace/FakeController.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Yii\View\Tests\Support\Controller\SubNamespace;
+namespace Yiisoft\Yii\View\Renderer\Tests\Support\Controller\SubNamespace;
 
 final class FakeController
 {