+++ tests/Support/Controllers/FakeCntrl.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Yii\View\Tests\Support\Controllers;
+namespace Yiisoft\Yii\View\Renderer\Tests\Support\Controllers;
 
 final class FakeCntrl
 {