+++ tests/Support/FakeCntrl.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Yii\View\Tests\Support;
+namespace Yiisoft\Yii\View\Renderer\Tests\Support;
 
 class FakeCntrl
 {