+++ tests/Support/FakeCsrfToken.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Yii\View\Tests\Support;
+namespace Yiisoft\Yii\View\Renderer\Tests\Support;
 
 use Yiisoft\Csrf\CsrfTokenInterface;
 use Yiisoft\Security\Random;