+++ tests/Support/NotCntrls/SubNamespace/FakeCntrl.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Yii\View\Tests\Support\NotCntrls\SubNamespace;
+namespace Yiisoft\Yii\View\Renderer\Tests\Support\NotCntrls\SubNamespace;
 
 final class FakeCntrl
 {